start migrating demo app to gorm-tools
