+++ examples/ag-demo/grails-app/dao/OrgRepo.groovy
@@ -1,13 +1,12 @@
 package agdemo
 
-import grails.plugin.dao.DaoMessage
-import grails.plugin.dao.DaoUtil
-import grails.plugin.dao.DomainException
-import grails.plugin.dao.GormDaoSupport
+import gorm.tools.repository.GormRepo
 import gorm.tools.beans.DateUtil
+import gorm.tools.repository.RepoMessage
+import gorm.tools.repository.RepoUtil
+import gorm.tools.repository.errors.EntityValidationException
 
-class OrgDao extends GormDaoSupport {
-    Class domainClass = Org
+class OrgRepo implements GormRepo<Org> {
 
     Map insert(params) {
         def org = new Org()
@@ -15,26 +14,25 @@ class OrgDao extends GormDaoSupport {
         persistWithParams(org, params)
 
 
-        [ok: true, entity: org, message: DaoMessage.created(org)]
+        [ok: true, entity: org, message: RepoMessage.created(org)]
     }
 
     Map update(params) {
         def org = Org.get(params.id.toLong())
         persistWithParams(org, params)
 
-        [ok: true, entity: org, message: DaoMessage.updated(org)]
+        [ok: true, entity: org, message: RepoMessage.updated(org)]
     }
 
     void persistWithParams(org, params) {
         if (params.addressDate && params.addressDate instanceof String) {
-          println params.addressDate
             params.addressDate = DateUtil.parseJsonDate(params.addressDate)
         }
         org.properties = params
         try {
-            save(org)
-            DaoUtil.flush()
-        } catch (DomainException e) {
+            org.persist()
+            RepoUtil.flush()
+        } catch (EntityValidationException e) {
             e.meta = [org: org]
             throw e
         }
