+++ gradle.properties
@@ -1,24 +1,11 @@
-grailsVersion=3.2.11
-gormVersion=6.0.12.RELEASE
-gradleWrapperVersion=3.4.1
-version=3.2.8-SNAPSHOT
+# grails/gradle version settings
+grailsVersion=3.3.10
+gormVersion=6.1.11.RELEASE
+gradleWrapperVersion=4.10.3
+# versions
+versionHibernate=5.2.17.Final
+versionBuildTestData=3.3.+
+versionViews=1.2.7
 
-# common
-title=Angle Grinder Grails Plugin
-description=Angular + Grails
-githubOrg=yakworks
-githubRepo=angle-grinder
 
-gormHibPluginVersion=6.1.8
-#buildTestDataVersion=3.3.0.RC2
-viewsVersion=1.2.6
-#hibernateVersion=5.2.12.Final
-
-# Bintray and/or artifactory / maven
-group=org.grails.plugins
-bintrayOrg=9ci
-bintrayRepo=grails-plugins
-
-# Gradle Settings
-#org.gradle.configureondemand=true This creates issues for bintrayUpload - commented out for now.
-#org.gradle.caching=true
+vPluginGormTools=6.1.11-v.9
