+++ plugin/grails-app/controllers/grinder/UrlMappings.groovy
@@ -1,4 +1,7 @@
-/* Copyright 2019. 9ci Inc. Licensed under the Apache License, Version 2.0 */
+/*
+* Copyright 2019 9ci Inc - Licensed under the Apache License, Version 2.0 (the "License")
+* You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+*/
 package grinder
 
 class UrlMappings {
