+++ plugin/grails-app/services/grinder/GrinderLabelService.groovy
@@ -1,8 +1,12 @@
-/* Copyright 2019. 9ci Inc. Licensed under the Apache License, Version 2.0 */
+/*
+* Copyright 2019 9ci Inc - Licensed under the Apache License, Version 2.0 (the "License")
+* You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+*/
 package grinder
 
 
 import javax.annotation.PostConstruct
+
 import grails.util.GrailsNameUtils
 
 //XXX - It doesnt seem to be used any more
