+++ version.properties
@@ -1,0 +1,3 @@
+version=3.2.8
+previousVersion=1.2.1
+snapshot=false
